# frontend/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"servers":[{"url":"/api"}],"paths":{"/auth/register":{"post":{"tags":["Authentication"],"summary":"Register User Endpoint","operationId":"register_user_endpoint_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/token":{"post":{"tags":["Authentication"],"summary":"Login For Access Token Endpoint","operationId":"login_for_access_token_endpoint_auth_token_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_for_access_token_endpoint_auth_token_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/users/me":{"get":{"tags":["Authentication"],"summary":"Read Users Me","operationId":"read_users_me_auth_users_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"put":{"tags":["Authentication"],"summary":"Update Users Me Endpoint","operationId":"update_users_me_endpoint_auth_users_me_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/auth/logout":{"post":{"tags":["Authentication"],"summary":"Logout User","description":"Деавторизация пользователя.\nДобавляет токен в черный список Redis.","operationId":"logout_user_auth_logout_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]},{"OAuth2PasswordBearer":[]}]}},"/report/draft/info":{"get":{"tags":["Reports"],"summary":"Get Draft Report Info Endpoint","description":"Получение ID и количества оценок уязвимости в черновике отчета текущего пользователя.","operationId":"get_draft_report_info_endpoint_report_draft_info_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssessmentBasketInfo"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/report/draft/status":{"get":{"tags":["Reports"],"summary":"Get Draft Report Status Endpoint","description":"Получение статуса активности и количества оценок уязвимости в черновике отчета текущего пользователя.","operationId":"get_draft_report_status_endpoint_report_draft_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DraftReportStatusInfo"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/report/draft/assessments":{"post":{"tags":["Reports"],"summary":"Add To Draft Report","description":"Добавление оценки уязвимости в черновик отчета (создает черновик, если его нет).","operationId":"add_to_draft_report_report_draft_assessments_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssessmentBasketItemAdd"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssessmentReportDetails"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/report/draft/assessments/{assessment_id}":{"delete":{"tags":["Reports"],"summary":"Remove From Draft Report","description":"Удаление оценки уязвимости из черновика отчета.","operationId":"remove_from_draft_report_report_draft_assessments__assessment_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"assessment_id","in":"path","required":true,"schema":{"type":"integer","title":"Assessment Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssessmentReportDetails"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Reports"],"summary":"Update Draft Report Item","description":"Изменение полей м-м (уровня защиты, комментария) для оценки уязвимости в черновике отчета.","operationId":"update_draft_report_item_report_draft_assessments__assessment_id__put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"assessment_id","in":"path","required":true,"schema":{"type":"integer","title":"Assessment Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssessmentBasketItemUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssessmentReportDetails"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/reports":{"get":{"tags":["Reports"],"summary":"Get Reports Endpoint","description":"Получение списка оформленных отчетов с фильтрацией.","operationId":"get_reports_endpoint_reports_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/ReportStatus"},{"type":"null"}],"title":"Status"}},{"name":"date_from","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Date From"}},{"name":"date_to","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Date To"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AssessmentReportSummary"},"title":"Response Get Reports Endpoint Reports Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/reports/{report_id}":{"get":{"tags":["Reports"],"summary":"Get Report Endpoint","description":"Получение детальной информации об отчете.","operationId":"get_report_endpoint_reports__report_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"report_id","in":"path","required":true,"schema":{"type":"integer","title":"Report Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssessmentReportDetails"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Reports"],"summary":"Update Report","description":"Изменение полей отчета (например, целевой системы).","operationId":"update_report_reports__report_id__put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"report_id","in":"path","required":true,"schema":{"type":"integer","title":"Report Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssessmentReportUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssessmentReportDetails"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Reports"],"summary":"Delete Report","description":"Удаление отчета-черновика (логическое).","operationId":"delete_report_reports__report_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"report_id","in":"path","required":true,"schema":{"type":"integer","title":"Report Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/reports/{report_id}/form":{"put":{"tags":["Reports"],"summary":"Form Report","description":"Сформировать отчет (создателем).","operationId":"form_report_reports__report_id__form_put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"report_id","in":"path","required":true,"schema":{"type":"integer","title":"Report Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssessmentReportFormPayload"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssessmentReportDetails"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/reports/{report_id}/complete":{"put":{"tags":["Reports"],"summary":"Complete Report","description":"Завершить отчет (модератором).","operationId":"complete_report_reports__report_id__complete_put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"report_id","in":"path","required":true,"schema":{"type":"integer","title":"Report Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssessmentReportDetails"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/reports/{report_id}/cancel":{"put":{"tags":["Reports"],"summary":"Cancel Report","description":"Отклонить отчет (модератором).","operationId":"cancel_report_reports__report_id__cancel_put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"report_id","in":"path","required":true,"schema":{"type":"integer","title":"Report Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssessmentReportDetails"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/vulnerabilities/":{"get":{"tags":["Vulnerabilities"],"summary":"Get Vulnerability Assessments","description":"Получение списка оценок уязвимости с возможностью фильтрации.","operationId":"get_vulnerability_assessments_vulnerabilities__get","parameters":[{"name":"title","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"}},{"name":"assessment_type","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/VulnerabilityAssessmentType"},{"type":"null"}],"title":"Assessment Type"}},{"name":"min_price","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Min Price"}},{"name":"max_price","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Max Price"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/VulnerabilityAssessment"},"title":"Response Get Vulnerability Assessments Vulnerabilities  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Vulnerabilities"],"summary":"Create Vulnerability Assessment","description":"Создание новой оценки уязвимости (только для модераторов).","operationId":"create_vulnerability_assessment_vulnerabilities__post","security":[{"OAuth2PasswordBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VulnerabilityAssessmentCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VulnerabilityAssessment"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/vulnerabilities/{assessment_id}":{"get":{"tags":["Vulnerabilities"],"summary":"Get Vulnerability Assessment","description":"Получение детальной информации об одной оценке уязвимости.","operationId":"get_vulnerability_assessment_vulnerabilities__assessment_id__get","parameters":[{"name":"assessment_id","in":"path","required":true,"schema":{"type":"integer","title":"Assessment Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VulnerabilityAssessment"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Vulnerabilities"],"summary":"Update Vulnerability Assessment","description":"Обновление существующей оценки уязвимости (только для модераторов).","operationId":"update_vulnerability_assessment_vulnerabilities__assessment_id__put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"assessment_id","in":"path","required":true,"schema":{"type":"integer","title":"Assessment Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VulnerabilityAssessmentUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VulnerabilityAssessment"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Vulnerabilities"],"summary":"Delete Vulnerability Assessment","description":"Удаление оценки уязвимости и связанного с ней изображения (только для модераторов).","operationId":"delete_vulnerability_assessment_vulnerabilities__assessment_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"assessment_id","in":"path","required":true,"schema":{"type":"integer","title":"Assessment Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/vulnerabilities/{assessment_id}/image":{"post":{"tags":["Vulnerabilities"],"summary":"Upload Image","description":"Загрузка/обновление изображения для оценки уязвимости (только для модераторов).","operationId":"upload_image_vulnerabilities__assessment_id__image_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"assessment_id","in":"path","required":true,"schema":{"type":"integer","title":"Assessment Id"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_image_vulnerabilities__assessment_id__image_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VulnerabilityAssessment"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AssessmentBasketInfo":{"properties":{"report_id":{"type":"integer","title":"Report Id"},"item_count":{"type":"integer","title":"Item Count"}},"type":"object","required":["report_id","item_count"],"title":"AssessmentBasketInfo"},"AssessmentBasketItemAdd":{"properties":{"assessment_id":{"type":"integer","title":"Assessment Id"}},"type":"object","required":["assessment_id"],"title":"AssessmentBasketItemAdd"},"AssessmentBasketItemUpdate":{"properties":{"protection_level":{"$ref":"#/components/schemas/ProtectionLevel"},"comment":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Comment"}},"type":"object","required":["protection_level"],"title":"AssessmentBasketItemUpdate"},"AssessmentComponent":{"properties":{"vulnerability_assessment":{"$ref":"#/components/schemas/VulnerabilityAssessment"},"protection_level":{"$ref":"#/components/schemas/ProtectionLevel"},"comment":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Comment"},"price_at_order_time":{"type":"string","title":"Price At Order Time"}},"type":"object","required":["vulnerability_assessment","protection_level","price_at_order_time"],"title":"AssessmentComponent"},"AssessmentReportDetails":{"properties":{"id":{"type":"integer","title":"Id"},"status":{"$ref":"#/components/schemas/ReportStatus"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"creator_login":{"type":"string","title":"Creator Login"},"created_by":{"type":"integer","title":"Created By"},"moderator_login":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Moderator Login"},"formation_date":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Formation Date"},"completion_date":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Completion Date"},"target_system_info":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Target System Info"},"risk_score":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Risk Score"},"components":{"items":{"$ref":"#/components/schemas/AssessmentComponent"},"type":"array","title":"Components","default":[]}},"type":"object","required":["id","status","created_at","creator_login","created_by"],"title":"AssessmentReportDetails"},"AssessmentReportFormPayload":{"properties":{"target_system_info":{"type":"string","minLength":5,"title":"Target System Info"}},"type":"object","required":["target_system_info"],"title":"AssessmentReportFormPayload"},"AssessmentReportSummary":{"properties":{"id":{"type":"integer","title":"Id"},"status":{"$ref":"#/components/schemas/ReportStatus"},"formation_date":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Formation Date"},"risk_score":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Risk Score"},"creator_login":{"type":"string","title":"Creator Login"}},"type":"object","required":["id","status","creator_login"],"title":"AssessmentReportSummary"},"AssessmentReportUpdate":{"properties":{"target_system_info":{"type":"string","minLength":5,"title":"Target System Info","default":""}},"type":"object","title":"AssessmentReportUpdate"},"AssessmentStatus":{"type":"string","enum":["deleted","available"],"title":"AssessmentStatus"},"Body_login_for_access_token_endpoint_auth_token_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","format":"password","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"format":"password","title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_for_access_token_endpoint_auth_token_post"},"Body_upload_image_vulnerabilities__assessment_id__image_post":{"properties":{"image":{"type":"string","format":"binary","title":"Image"}},"type":"object","required":["image"],"title":"Body_upload_image_vulnerabilities__assessment_id__image_post"},"DraftReportStatusInfo":{"properties":{"is_active":{"type":"boolean","title":"Is Active"},"item_count":{"type":"integer","title":"Item Count"}},"type":"object","required":["is_active","item_count"],"title":"DraftReportStatusInfo"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ProtectionLevel":{"type":"string","enum":["none","basic","full"],"title":"ProtectionLevel"},"ReportStatus":{"type":"string","enum":["draft","deleted","formed","completed","cancelled"],"title":"ReportStatus"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token"],"title":"Token"},"UserCreate":{"properties":{"login":{"type":"string","minLength":3,"title":"Login"},"password":{"type":"string","minLength":4,"title":"Password"}},"type":"object","required":["login","password"],"title":"UserCreate"},"UserRead":{"properties":{"id":{"type":"integer","title":"Id"},"login":{"type":"string","title":"Login"},"is_moderator":{"type":"boolean","title":"Is Moderator"}},"type":"object","required":["id","login","is_moderator"],"title":"UserRead"},"UserUpdate":{"properties":{"login":{"anyOf":[{"type":"string","minLength":3},{"type":"null"}],"title":"Login"}},"type":"object","title":"UserUpdate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VulnerabilityAssessment":{"properties":{"title":{"type":"string","maxLength":128,"minLength":3,"title":"Title"},"short_description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Short Description"},"description":{"type":"string","title":"Description"},"price":{"type":"string","title":"Price"},"impact_level":{"type":"integer","maximum":3.0,"minimum":1.0,"title":"Impact Level"},"assessment_type":{"$ref":"#/components/schemas/VulnerabilityAssessmentType"},"id":{"type":"integer","title":"Id"},"status":{"$ref":"#/components/schemas/AssessmentStatus"},"image_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image Url"}},"type":"object","required":["title","description","price","impact_level","assessment_type","id","status"],"title":"VulnerabilityAssessment"},"VulnerabilityAssessmentCreate":{"properties":{"title":{"type":"string","maxLength":128,"minLength":3,"title":"Title"},"short_description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Short Description"},"description":{"type":"string","title":"Description"},"price":{"anyOf":[{"type":"number","exclusiveMinimum":0.0},{"type":"string"}],"title":"Price"},"impact_level":{"type":"integer","maximum":3.0,"minimum":1.0,"title":"Impact Level"},"assessment_type":{"$ref":"#/components/schemas/VulnerabilityAssessmentType"}},"type":"object","required":["title","description","price","impact_level","assessment_type"],"title":"VulnerabilityAssessmentCreate"},"VulnerabilityAssessmentType":{"type":"string","enum":["network_scan","web_app_pentest","infrastructure_audit"],"title":"VulnerabilityAssessmentType"},"VulnerabilityAssessmentUpdate":{"properties":{"title":{"anyOf":[{"type":"string","maxLength":128,"minLength":3},{"type":"null"}],"title":"Title"},"short_description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Short Description"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"price":{"anyOf":[{"type":"number","exclusiveMinimum":0.0},{"type":"string"},{"type":"null"}],"title":"Price"},"impact_level":{"anyOf":[{"type":"integer","maximum":3.0,"minimum":1.0},{"type":"null"}],"title":"Impact Level"},"assessment_type":{"anyOf":[{"$ref":"#/components/schemas/VulnerabilityAssessmentType"},{"type":"null"}]}},"type":"object","title":"VulnerabilityAssessmentUpdate"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/auth/token"}}}}}}